refactor: extract teams management to dedicated /teams route
